Support patch/diff files

+++ make.rb
@@ -11,27 +11,27 @@ AUTHOR = 'Yukihiro Matsumoto <matz@ruby-lang.org>'
 FILES = <<EOD
 ruby-0.49.tar.gz 214390 2015-07-25T01:35:23.000Z
 ruby-0.50.tar.gz 219673 2015-07-25T01:35:23.000Z
-ruby-0.51-0.52.diff.gz 6334 2018-01-25T04:29:11.000Z
 ruby-0.51.tar.gz 236604 2015-07-25T01:35:23.000Z
+ruby-0.51-0.52.diff.gz 6334 2018-01-25T04:29:11.000Z
 ruby-0.54.tar.gz 241736 2015-07-25T01:35:24.000Z
-ruby-0.55-0.56.diff.gz 5644 2018-01-25T04:29:11.000Z
 ruby-0.55.tar.gz 242261 2015-07-25T01:35:24.000Z
+ruby-0.55-0.56.diff.gz 5644 2018-01-25T04:29:11.000Z
 ruby-0.60.tar.gz 223532 2015-07-25T01:35:24.000Z
 ruby-0.62.tar.gz 224325 2018-06-13T06:51:19.000Z
 ruby-0.62.tar.gz.broken 238311 2018-06-13T06:51:19.000Z
 ruby-0.63.tar.gz 224525 2018-06-13T06:51:19.000Z
 ruby-0.63.tar.gz.broken 238511 2018-06-13T06:51:19.000Z
 ruby-0.64.tar.gz 227088 2015-07-25T01:35:28.000Z
-ruby-0.65-0.66.diff.gz 5020 2018-01-25T04:29:11.000Z
 ruby-0.65.tar.gz 237040 2015-07-25T01:35:27.000Z
+ruby-0.65-0.66.diff.gz 5020 2018-01-25T04:29:11.000Z
 ruby-0.66-0.67.diff.gz 17897 2018-01-25T04:29:10.000Z
 ruby-0.67-0.68.diff.gz 17074 2018-01-25T04:29:11.000Z
 ruby-0.69.tar.gz 228241 2015-07-25T01:35:28.000Z
 ruby-0.70-patch 2442 2018-01-25T04:29:11.000Z
-ruby-0.71-0.72.diff.gz 4785 2018-01-25T04:29:11.000Z
 ruby-0.71.tar.gz 226276 2015-07-25T01:35:28.000Z
-ruby-0.73-950413.tar.gz 237953 2015-07-25T01:35:28.000Z
+ruby-0.71-0.72.diff.gz 4785 2018-01-25T04:29:11.000Z
 ruby-0.73.tar.gz 228960 2015-07-25T01:35:28.000Z
+ruby-0.73-950413.tar.gz 237953 2015-07-25T01:35:28.000Z
 ruby-0.76.tar.gz 243351 2015-07-25T01:35:28.000Z
 ruby-0.95.tar.gz 311119 2015-07-25T01:35:29.000Z
 ruby-0.99.4-961224.tar.gz 352090 2015-07-25T01:35:29.000Z
@@ -49,8 +49,6 @@ EOD
 
 files = FILES.lines.reject{|line|
   line =~ /broken/
-}.reject{|line|
-  line =~ /diff|patch/  # Pull Request please.
 }.map{|line|
   line.split.first
 }
@@ -80,13 +78,25 @@ Dir.chdir("#{out_dir}") do
   files.each do |filename|
     tar_path = "archives/#{filename}"
     puts tar_path
-    system *%W"tar -xzf #{tar_path}", exception: true
-    FileUtils.rm_rf(Dir.glob("repo/*"))
-    Dir.glob(%w"ruby/* ruby-*/*") do |n|
-      File.rename(n, "repo/#{File.basename(n)}")
+    case filename
+    when /\Aruby-[\d.]+(-[\d.]+)\.diff\.gz\z/
+      message = "ruby#{$1}"
+      IO.popen(%W[gzcat #{tar_path}], "r") do |gz|
+        system *%W"patch -d repo -p1", in: gz, exception: true
+      end
+    when /-patch\z/
+      message = $`
+      system *%W"patch -d repo -p1", in: tar_path, exception: true
+    when /\.tar\./
+      message = $`
+      system *%W"tar -xzf #{tar_path}", exception: true
+      FileUtils.rm_rf(Dir.glob("repo/*"))
+      Dir.glob(%w"ruby/* ruby-*/*") do |n|
+        File.rename(n, "repo/#{File.basename(n)}")
+      end
+      FileUtils.rmdir(Dir.glob(%w"ruby ruby-*"))
     end
-    FileUtils.rmdir(Dir.glob(%w"ruby ruby-*"))
     system *%W"git -C repo add .", exception: true
-    system *%W"git -C repo commit --author=#{AUTHOR} -m #{filename}", exception: true
+    system *%W"git -C repo commit --author=#{AUTHOR} -m #{message}", exception: true
   end
 end
